Remove adaptee class
